Add handling space subdomain (#7)
When an app runs in a space's iframe on Cobot and that space has a custom domain, OAuth needs to use the custom domain and not www.cobot.me, otherwise users have to log in again when accessing the app. This is now accomplised by passing the space's subdomain to the `/auth/cobot` endpoint.